Remove "as first status"
This makes the sentence confusing. 
An alternate way of writing it if you wish to keep that bit in: "The first status used when developing a new API endpoint." I did not change it to this because it would not be consistent with how the rest of the items on the list were written.